SourceDocument now has a simplify method that discards typing events which are not a member of the list made by playOutEvents and then 'homogenizes' the ones left over (this means that they are assigned a new time which is related to their position in the document in order to counter the effect of absence of discarded events). The shuffledEventsTest method has now been changed to also include a test of the simplify method. It has now been renamed shuffledAndSimplificationTest().
Length test: Because I slightly changed the generateEvents method a while ago (which is only used for testing)
the length test was failing.
The first generated event has time = startTime. In the test routine the '>' symbol was being added at the same time
as the first character, when actually it should of been
[{text = '<', time = 0} , {text = '>', time = 1}, { generateEvents starting from time = 2 }]

All tests, including the test for the new simplify feature, now pass.