Fell asleap in the afternoon... Now I'm paying the price.
A couple of things that haven't been tried out because an old version of the bot
is running somewhere right now, so I've included a backdoor routine to shut the
bot down once the existing one is shut down.

Also put something that might stop the carets jumping about when two people are
editing at the same time but I haven't tried it yet.