In the process of trying to make sure document closures were marked properly I noticed that on restart the LoginUI seemed to be running another instance of itself inside itself, so probably when the window closed all the data in its fields were not being thrown away. In the end I found the only solution I could think of to this problem was to have another object, which I called CiderApplication, to control the starting and restarting of loginUI.
The entry point for the application is now located in CiderApplication.
The first time you try to run this eclipse might not recognise the new entry
point, so you may want to right-click on CiderApplication then go to
Run-as -> Java Application.

Anyway it turns out that Swing will call windowClosed or windowClosing depending
on whether you close it with the x button or not. I also made a couple of
cosmetic changes to the history viewer.